Constants and Function Extension

+++ src/model/Item.kt
@@ -1,8 +1,10 @@
 package model
 
+import java.math.BigDecimal
+
 class Item(
     var name: String,
-    var price: Double
+    var price: BigDecimal
 ) {
 
 }
